Updates to DuplicateRemover
Instead of showing duplicates in a listbox, they are now saved to a file
that will open up. Added a delete button so the user can check over the
files that will be deleted before deleting them all. Added a browse
button to let the user get to the directory instead of only typing it by
hand. Roms in subdirectories will be analyzed instead of just the root
folder. The regions textbox is actually used for checking regions now.