Skeleton for batch process
An immediate problem is that there will need to be a lot of code changes to make it that a form pops up, the user goes through it, and then when they submit that form, once everyone is done it will then just link them to the next form.

The "easier" solution is to just make new forms that just reference the code on the older forms that will continue linking.

The "harder" solution is to somehow thread out the processes to create linking, but forms can only be loaded on the main thread.